Report if products are gotten from Event, Lumi or Run without corresponding consumes call
Each time a data product is gotten from the Event, LuminosityBlock or Run without having the module calling the corresponding consumes, mayConsume or consumesMany call we issue a MessageLogger warning.
This required passing back from the Principal exactly which ProductHolderIndex was used to find the data. In the fully threaded framework this will be changed to throw an exception.