Sources can tell the framework they need to synchronize
Added the potential to allow a Source to tell the framework it needs the
framework to finish any outstanding concurrent work before the Source
can proceed. This allows the Source to handle any cases where it can't
work in parallel. For now, the EventProcessor just ignores the statement
since everything is singled threaded right now.