Properly preallocate all per module per stream items
Percolated the PreallocationConfiguration into the module base classes and passed the number of streams to the appropriate virtual functions which are overwritten in the appropriate inheriting classes. The per Run and LuminosityBlock items are still just single items since having multiple such transitions is a later phase of the project.