SubProcess has own copy of ProcessHistoryRegistry
Now has its own copy of the ProcessHistoryRegistry as preparation for thread-safety. As part of the change, now get the parent’s reducedProcessHistoryID from the RunPrincipal and not from the parent ProcessHistoryRegistry (which is no longer accessible).
Full thread-safety will require each Stream to have its own ProcessHistoryRegistry and HistoryAppender.