Changed internals of BranchIDListHelper to allow fixBranchListIndexes to be const
The structure which is used to convert the BranchListIndex read from the source to the indexes used in the job was converted from a map to a vector. This allows calls to fixBranchListIndexes to be immutable.  However, RootFile was depending on SecondarySources to get dummy values for the BranchListIndexes which came from calling map<>::operator[]. RootFile was changed so it would instead properly initialize the BranchListIndexes by calling ‘updateFromInput on a SecondarySource.
In addition, comments were added to clarify how this data structure is used in the threaded framework.