Moved knowledge of which BranchListIndex corresponds to any products produced in this Event from ProductRegistry to BranchIDListHelper
The BranchIDListHelper was already being used to determine the produced BranchListIndex and the feeding that back into the transient info held by ProductRegistry. That information was then only used in EventPrincipal::fillEventPrincipal which already has access to the BranchIDListHelper. Since the code no longer modifies the ProductRegistry, the member function was changed from ‘updateRegistries’ to ‘updateFromRegistry’.
Many unit tests were updated to match the new interface.