RandomNumberGenerator interface
This change adds to the RandomNumberGenerator interface
to allow clients to migrate to use the interface
that will be used in the multithreaded version of the
service. This is not the multithreaded version of the
service. With this change the service still cannot be
used in a job where the Framework is set to use multiple
streams or multiple concurrent luminosity blocks.