Move code shared by executable and plugin to an existing library
The unit test analyzer TestBeginEndJobAnalyzer has a static used
by the unit test to see that all transitions happened. When compiled
with clang the unit test failed presumably because we had mulitple
copies of the static. This change fixes the problem and makes the code
obey the C++ one definition rule.