Bug fix for NewEventStreamFileReader
Make the inputFileTransitionsEachEvent option work
with NewEventStreamFileReader. This was accidently
broken in 2012, although as far as I know no one
normally uses this option with this source. It was
noticed when someone used the option to debug a
related problem in another source. Added a unit test
so we will notice if this problem recurs.

The bug caused the input file to be prematurely closed
when the fake input file transitions were occurring.