Move HolderToVectorTrait specializations to own files
To avoid unnecessary clutter in header files and to allow faster compilation for the case where the specialization isn't needed, the specializations of HolderToVectorTrait were moved to their own files.
The use of the other header's include guards to guarantee we only include the code when the other two files are included is unorthodox but works in this case since all files live in the same package.