Ref cache points at element not container
Makes the pointer cached in a Ref point to the
element in the container instead of pointing
to the container. Removes the product function
from classes related to Ref, RefToBase, Ptr, and
vectors of them. Remove constructor of Ref from
RefVector. Add Ref constructor that takes ProductID,
pointer to element, and a key. Remove RefProd
constructor from Ref and RefProd constructor from
RefVector. Remove  RefToBaseProd constructors from
Ref, RefToBase, and View. Make RefToBaseProd from
RefProd constructor explicit.

This is all motivated by changes to the View class
that indirectly require this to work properly.

Many changes to make code that depends on these changes
continue work as before.

Additions to make it easier to build AssociationMap
with a one argument constructor.

A previous commit required the AssociationMap be
initialized with RefProd before calling insert. This
commit fixes some but not all of the cases where
this caused failures.

Note this is work in progress. If you try to run with
this commit failures will occur.