Integrate CMSSW with HTCondor's update service.
This commit provides a new default service, CondorStatusService,
which automatically reports basic progress statistics (# events,

The service will automatically detect if it's running as part of
a HTCondor job and probes to see if HTCondor has user-level updates
enabled (and is a sufficient version).  If either check fails, then
the service does not register any callbacks with the framework and
effectively becomes a no-op.

To see if it is within a HTCondor job, the service looks for the
_CONDOR_CHIRP_CONFIG environment variable - a very cheap check.
To see if HTCondor supports this feature, it spawns a new process
and looks at the exit code (more expensive).

This uses the 'set_job_attr_delayed' mechanism of HTCondor,
which causes these updates to 'tag along' with the existing updates
for memory / disk / CPU usage.  Hence, the extra cost is the few
extra bytes in the update packet that goes out once every 5 minutes.
The update only goes as far as the local daemon; condor_chirp does
not wait for it to propagate to the remote host.  Hence, it exits
rapidly.

CMSSW will only update once every updateIntervalSeconds (defaults
to 15 minutes).  The HTCondor worker node and central components
all have additional rate limiting mechanisms to prevent overload.

(cherry picked from commit 17a9f627fab4df84ceffa82454f73aca4eec61e5)