Remove function getProcessParameterSet
Remove the function getProcessParameterSet
from CMSSW. Previous PRs removed most of the
places where this function was used. This
PR gets a few minor uses that were left in
the Framework code. Also removes setProcessParameterSetID
and getProcessParameterSetID.

This function is being removed because it does not work
with SubProcesses and caused threading issues requiring
the use of a mutex.