Remove module circular dependency check from EventPrincipal
We previously checked that a module wasn't indirectly calling itself in unscheduled via a run time check. With the requirements on consumes calls, we now check for circular dependencies before running the job. In addition, this implementation would not be thread safe. However, if we ever did need to add back a runtime check we could use the Context information instead.