Improve const correctness of Principal internals
Reduced the number of mutable data members used.
Switched to using propagate_const<> to make pointer like data members have deep const behavior.
Changes the API of a number of internal classes to be properly const or non-const.
These changes are meant to make it easier to reason about the thread-safety of the classes.