Removed Transient from ProductProvenance
Transient was primarily used to do a cache of edm::Parentage even though the information is stored in ParentageRegistry. The Parentage object is only requested on output and only once per branch so the cache doesn't look to be all that useful. The cache is sometimes updated lazily which causes thread-safety issues. The easiest solution is just to remove the use of the Transient all together.
(cherry picked from commit dc7822b82b534cb3cde321c0a809e2670cf3fd0a)