Added Modifier.toReplaceWith
Added the method toReplaceWith to Modifier. This function is used to replicate the state of the fromObj into the existing toObj. This allows a larger conversion of the toObj than can be handled to the toModify method.
(cherry picked from commit c180ba7ca82909ed3e737a6dad131ab064ba8ec7)