Only use one thread for the test
The default for TBB is to use as many threads as cores. This is clearly unnecessary for this case so instead we just set it to one thread.
This test would infrequently fail with a strange message about missing file descriptor. I believe this was a byproduct of using so many threads.