PuttableProductResolver gives wait task to Module
In order to handle the case where a module does not put a data
product, e.g. because of an exception, we still need to release
any modules waiting on the data product. We accomplish this by
having the PuttableProductResolver attach a waiting task to the
module to know when it has completed.