Fix randomly failing framework unit test
The test run_global_stream_one.sh includes a test to check that
when using modules which share a resource that only one module
runs at a time. Previously, the measurement mechanism also included
TriggerResults and the source in the timing which could create
false positives and fail the test. Now the test excludes those
two from the measurement so the test should never see more than one
module running at a time.