Pass thread count to gridpack run scripts.
Altered the ExternalLHEProducer to pass information from the cmsRun
process to the script that launches the generators.  This requires
a small change to EDProducerBase so producers are informed of the
number of threads in the current configuration.

Generators that are ready to run multi-process / multi-threaded
will use this extra argument; otherwise, this is ignored.