Make MessageLogger give consistent results
After allowing the beginStreamLuminosityBlock for each stream to run on different threads, this caused the per threads MessageDrop::runEvent used during the call to InputSource::issueReports to be set to a value based on what the thread had previously done. That value could change job to job. To make it consistent, the MessageDrop::runEvent is set to a specific value before that call will be made.

To help track down problems with Global or Stream transitions, the MessageLogger now clearly distinguished between the two.