Enforce processing only one transition at a time
Use a SerialTaskQueue to enforce processing of only one Run or LuminosityBlock transition at a time for module types which need the constraint.
Begin transition items are placed in the proper queue and then that queue gets paused when the task runs. The Queue does not resume until the
corresponding end transition runs. This then frees any waiting begin transitions to start.