Add ability to set exception early in waiting tasks
Allow an exception to be specified early to the held task. This helps with the case where an exception has happened but the task cannot be release yet. However, as part of the exception clean up, new tasks might be run which also fail. By setting an exception early, we can guarantee that the progenitor exception is the one reported out of the framework.