Add ability to validate user plugins
Extended the ParameterSet validation system to support the case where user plugins have different parameters. The code can now use a `fillPSetDescription` static member function of a plugin to validate the configuration before the module using the plugin is constructed. See the documentation in  FWCore/ParameterSet/interface/PluginDescription for the full details.