Fix race condition in EventProcessor
First note that that vast majority of the time,
the correct code path wins the data race and there
is no problem, almost always. That is why we have
rarely seen failures related to this.

The condition is related to the deletion of the
LuminosityBlockProcessingStatus. If we want to
call endRun after exiting processLumis, then
this status object must have already been destroyed
because it holds a shared_ptr to the RunResources
object that prevents endRun from running. If it isn't
destroyed then endRun may get called at some random
time much later which causes all kinds of problems. Seg
faults would be likely.