Simplify globalEndLumiAsync
Simply the function globalEndLumiAsync so it is more clear
to someone reading or maintaining the code who owns the lumi
status object. The status object is now moved into the first
task instead of copied. It is no longer moved into the second
task, which never really needed the shared_ptr in the first place.

I also edited the comments a bit and renamed a few things
to make this part of the code more clear.

This commit should not change behavior in any way.