Delete unscheduled modules whose products are not consumed by anything
The corresponding WorkerT continues to exist, but is removed from all
WorkerManagers so that it does not get called for any transitions. The
deleted modules are removed from PathsAndConsumesOfModules.

Allow worker_ to be nullptr in UnscheduledProductResolver, can happen
if a module is deleted early.

The Worker::description() returns now nullptr if module was deleted.
The description() is not expected to be called when the module is
deleted, so this change is mostly a safety check in the hope of
getting a segfault if the nullptr is accessed.