Make empty and '*' debugModules to result in all debug messages to be printed
Previously there was no way to enable debug messages for non-module
code. But we want to be able to issue debug messages e.g. from
Services or framework. In addition, the empty debugModules (that is
also the default value) leading to no debug messages being issued
seemed now a poor choice (given the amount of complaints in the past
on how difficult MessageLogger is to configure, especially for debug
messages).